Moved GEP into a subpackage and made minor changes to chromosome factory
Refactor: the maximum arity is now tracked in the ChromosomeFactory superclass.
Refactor: GeneExpressionProgramming has been moved to the ea.strategies package so it doesn't get bloated as we add more algorithms.
Documentation:  added a very small amount of documentation to ChromosomeFactory.
