+++ api/src/main/java/org/iconic/ea/chromosome/ChromosomeFactory.java
@@ -7,7 +7,7 @@ import java.util.List;
 
 /**
  * <p>
- * A ChromosomeFactory is used for easy replication of chromosomes.
+ * A ChromosomeFactory is used for easily replicating chromosomes.
  * </p>
  *
  * @param <T> The type of chromosome to be constructed
@@ -15,18 +15,22 @@ import java.util.List;
  */
 public abstract class ChromosomeFactory<T extends Chromosome<R>, R> {
     protected final List<FunctionalPrimitive<R, R>> functionalPrimitives;
+    private int maxArity;
 
     protected ChromosomeFactory() {
         functionalPrimitives = new LinkedList<>();
+        this.maxArity = 0;
     }
 
+    /**
+     * <p>Returns a new chromosome constructed according to the parameters of this factory</p>
+     *
+     * @return a chromosome constructed by the factory
+     */
     public abstract T getChromosome();
 
     public List<FunctionalPrimitive<R, R>> getFunctionalPrimitives() {
-        List<FunctionalPrimitive<R, R>> clone = new LinkedList<>();
-        clone.addAll(functionalPrimitives);
-
-        return clone;
+        return functionalPrimitives;
     }
 
     public FunctionalPrimitive<R, R> getFunction(int i) {
@@ -34,7 +38,26 @@ public abstract class ChromosomeFactory<T extends Chromosome<R>, R> {
     }
 
     public void addFunction(List<FunctionalPrimitive<R, R>> functions) {
-        functionalPrimitives.addAll(functions);
+        assert (functions.size() > 0);
+
+        getFunctionalPrimitives().addAll(functions);
     }
 
+    /**
+     * <p>Returns the maximum arity of the functions used by this factory</p>
+     *
+     * @return the maximum arity of the functions used by the factory
+     */
+    public int getMaxArity() {
+        return maxArity;
+    }
+
+    /**
+     *  <p>Sets the maximum arity of the functions used by this factory to the provided value</p>
+     *
+     * @param maxArity the new maximum of the functions used by the factory
+     */
+    protected void setMaxArity(int maxArity) {
+        this.maxArity = maxArity;
+    }
 }
